+++ src/map.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const map = new WeakMap();
 
 module.exports = map;